Started working on the ManageOpportunities Side of the project
Started working on the ManageOpportunities Side of the project
8fcb94d
Got the following methods working:
-Create
-delete
-ManageOpportunities

currently working on:
Edit (Have not checked if changes are saved - they should. And Also the DateTime is not bennif put int the input field for Time of Event.)

Added all of the required backend for the IOpportunityRepository... now Just need to fill in the methods.